Fix import cycle between models and tokenization
The cycle is fixed by moving `remove_eos_bos` to `curated_transformers.models`.
Since this layer operates on transformer output, it seems like a better location
anyway.
